REFSANS: Fix type conversion
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/25734
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>

+++ nicos_mlz/refsans/gui/monitorwidgets.py
@@ -271,7 +271,7 @@ class TimeDistance(NicosWidget, TimeDistanceWidget):
                 elif devname == 'disc2_pos':
                     self._disk2_pos = int(value)
                 elif devname == 'fp':
-                    self._fp = int(value)
+                    self._fp = float(value)
             elif param == 'phase':
                 if devname.startswith('chopper'):
                     if value is not None:
